dom/indexedDB/test/test_wrappedArray.xhtml

<?xml version="1.0"?>
<?xml-stylesheet type="text/css" href="chrome://global/skin"?>
<?xml-stylesheet type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css"?>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1509168
-->
<window title="Mozilla Bug 1509168"
        xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul">
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"/>

  <!-- test results are displayed in the html:body -->
  <body xmlns="http://www.w3.org/1999/xhtml">
  <a href="https://bugzilla.mozilla.org/show_bug.cgi?id=1509168"
     target="_blank">Mozilla Bug 1509168</a>
  </body>

  <!-- test code goes here -->
  <script type="application/javascript">
  <![CDATA[
  function go() {
    window.iwin = document.getElementById('ifr').contentWindow;

    let evilarr = new iwin.Uint8ClampedArray(3);

    try {
      window.indexedDB.cmp(evilarr, undefined);
    } catch (e) {}

    ok(true, "should reach here instead of crash");
  }
  ]]>
  </script>
  <iframe id="ifr" onload="go();" />
</window>
